// templates/base/src/common/enums/permission.enum.ts
export enum PermissionEnum {
  // EMPLOYEE
  EMPLOYEE_CREATE = 'employee.create',
  EMPLOYEE_READ = 'employee.read',
  EMPLOYEE_UPDATE = 'employee.update',
  EMPLOYEE_DELETE = 'employee.delete',

  // DEPARTMENT
  DEPARTMENT_CREATE = 'department.create',
  DEPARTMENT_READ = 'department.read',
  DEPARTMENT_UPDATE = 'department.update',
  DEPARTMENT_DELETE = 'department.delete',

  // DESIGNATION
  DESIGNATION_CREATE = 'designation.create',
  DESIGNATION_READ = 'designation.read',
  DESIGNATION_UPDATE = 'designation.update',
  DESIGNATION_DELETE = 'designation.delete',


  //Attendence

  ATTENDANCE_CREATE = 'attendance.create',
  ATTENDANCE_READ = 'attendance.read',
  ATTENDANCE_UPDATE = 'attendance.update',
  ATTENDANCE_DELETE = 'attendance.delete',

  EMPLOYEE_DASHBOARD_READ = 'dashboard.read',

  HR_DASHBOARD_READ = 'hr_dashboard.read',

  ATTENDANCE_CORRECTION_CREATE = 'attendence_correction.create',
  ATTENDANCE_CORRECTION_READ = 'attendence_correction.read',
  ATTENDANCE_CORRECTION_UPDATE = 'attendence_correction.update',
  ATTENDANCE_CORRECTION_DELETE= 'attendence_correction.delete',

  LEAVE_CREATE='leave.create',
  LEAVE_READ='leave.read',
  LEAVE_UPDATE='leave.update',
  LEAVE_DELETE='leave.delete',

  WEEKEND_CREATE='weekend.create',
  WEEKEND_READ='weekend.read',
  WEEKEND_UPDATE='weekend.update',
  WEEKEND_DELETE='weekend.delete',

  HOLIDAY_CREATE='holiday.create',
  HOLIDAY_READ='holiday.read',
  HOLIDAY_UPDATE='holiday.update',
  HOLIDAY_DELETE='holiday.delete',


}